pyproject.toml
[project]
name = "xo1997-gallery"
version = "0.1.0"
description = "xo1997 画廊 - AI 智能助手与组织架构生成工具"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6.0",
    "PySideSix-Frameless-Window>=0.8.0",
    "darkdetect",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-ollama>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.2.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "core", "data"]
